refactor: resolve sonar complexity and length violations in package modules
Part of turning the dev Sonar quality gate green (raes-strict fails on any
new-code violation). Fifteen recently-touched package modules each carried
one violation: two files exactly one line over the 500-line cap (S104),
four functions over 100 lines (S138), and nine functions with cyclomatic
complexity 11-13 against the 10 threshold.

All changes are behavior-preserving extractions with no public API,
contract, or schema change:

- raes/composition/_expand.py: move per-import provenance assembly into
  _import_provenance_additions
- raes/phase_contracts.py: hoist the two constraint-pointer shape
  predicates out of _validate_constraint
- raes_backend_libvirt/capability_envelope.py: split the operating-system
  identity probe out of _out_of_envelope_terms
- raes_backend_protocols/provisioner_capabilities.py: extract
  operating-system row and artifact-kind validation helpers
- raes_backend_stubs/stubs.py: extract the operations fold from apply
- raes_conformance/snapshot_semantics.py: extract entry and disclosure
  builders from _snapshot_from_envelope
- raes_contracts/apparatus.py, contracts/capabilities.py,
  realization_envelope_carrier.py: extract validator helper functions
- raes_contracts/contracts/bundle.py: move _runtime_schema_bundle and its
  imports into contracts/bundle_runtime.py (bundle output verified
  identical by tools/check_generated_schemas.py)
- raes_processor/planner/operations.py: extract ordered apply/delete op
  builders
- raes_processor/semantics/realization_runtime_evaluation.py: hoist the
  observation-corroboration predicate
- raes_runtime/backend_calls.py: move the ApplyResult shape validators
  into backend_call_contracts.py
- raes_runtime/control_plane_api_participant_retrieval.py: fold the three
  governed-view routes onto one shared resolve/audit flow
- raes_runtime/control_plane_store.py: move snapshot payload builders into
  control_plane_store_payloads.py (also keeps the file under the repo's
  own 500-line policy cap)

Verification: nox -s tests (7,005 passed, coverage gate green), ruff
format and lint clean, tools/check_generated_schemas.py identical,
tools/check_repo_policy.py pass.

Closes #1153
