+++ implementations/python/packages/raes_runtime/backend_calls.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from collections.abc import Callable, Iterable
+from collections.abc import Callable
 from copy import deepcopy
 
 from raes_contracts.addressing import require_compiled_address
@@ -21,6 +21,11 @@ from .backend_account_credentials import (
     plan_arguments_have_account_credentials,
     sanitize_account_credential_result,
     value_free_backend_diagnostics,
+)
+from .backend_call_contracts import (
+    _apply_result_contract_violation,
+    _diagnostics_iterable_violation,
+    _diagnostics_values_violation,
 )
 from .backend_realization_authority import (
     _apply_authority_diagnostics,
@@ -325,73 +330,6 @@ def _backend_contract_invalid(address: str, message: str) -> Diagnostic:
 
 def _failed_apply_result(snapshot: RuntimeSnapshot, diagnostic: Diagnostic) -> ApplyResult:
     return ApplyResult(success=False, snapshot=snapshot, diagnostics=[diagnostic])
-
-
-def _diagnostics_iterable_violation(result: object, address: str) -> str | None:
-    message = None
-    if not isinstance(result, Iterable) or isinstance(result, (str, bytes)):
-        message = f"Backend method '{address}' returned {type(result).__name__}; expected diagnostics iterable."
-    return message
-
-
-def _diagnostics_values_violation(diagnostics: list[object], address: str) -> str | None:
-    message = None
-    if any(not isinstance(diagnostic, Diagnostic) for diagnostic in diagnostics):
-        message = f"Backend method '{address}' returned a diagnostics iterable containing non-Diagnostic values."
-    return message
-
-
-def _apply_result_contract_violation(result: object, address: str) -> str | None:
-    message = _apply_result_shape_violation(result, address)
-    if message is None and isinstance(result, ApplyResult):
-        message = _apply_result_diagnostics_violation(result, address)
-    if message is None and isinstance(result, ApplyResult):
-        message = _apply_result_changed_addresses_violation(result, address)
-    if message is None and isinstance(result, ApplyResult):
-        message = _apply_result_details_violation(result, address)
-    return message
-
-
-def _apply_result_shape_violation(result: object, address: str) -> str | None:
-    message = None
-    if not isinstance(result, ApplyResult):
-        message = f"Backend method '{address}' returned {type(result).__name__}; expected ApplyResult."
-    elif not isinstance(result.snapshot, RuntimeSnapshot):
-        message = (
-            f"Backend method '{address}' returned ApplyResult.snapshot "
-            f"as {type(result.snapshot).__name__}; expected RuntimeSnapshot."
-        )
-    return message
-
-
-def _apply_result_diagnostics_violation(result: ApplyResult, address: str) -> str | None:
-    message = None
-    if not isinstance(result.diagnostics, Iterable) or isinstance(result.diagnostics, (str, bytes)):
-        message = (
-            f"Backend method '{address}' returned ApplyResult.diagnostics "
-            f"as {type(result.diagnostics).__name__}; expected iterable."
-        )
-    elif any(not isinstance(diagnostic, Diagnostic) for diagnostic in result.diagnostics):
-        message = f"Backend method '{address}' returned ApplyResult.diagnostics containing non-Diagnostic values."
-    return message
-
-
-def _apply_result_changed_addresses_violation(result: ApplyResult, address: str) -> str | None:
-    message = None
-    if not isinstance(result.changed_addresses, list):
-        message = (
-            f"Backend method '{address}' returned ApplyResult.changed_addresses "
-            f"as {type(result.changed_addresses).__name__}; expected list."
-        )
-    elif any(not isinstance(changed_address, str) for changed_address in result.changed_addresses):
-        message = f"Backend method '{address}' returned ApplyResult.changed_addresses containing non-string values."
-    return message
-
-
-def _apply_result_details_violation(result: ApplyResult, address: str) -> str | None:
-    if isinstance(result.details, dict):
-        return None
-    return f"Backend method '{address}' returned ApplyResult.details as {type(result.details).__name__}; expected dict."
 
 
 def _snapshot_contract_diagnostics(
