+++ implementations/python/packages/raes_runtime/control_plane_store.py
@@ -9,10 +9,6 @@ from dataclasses import dataclass, field
 from enum import Enum
 from typing import TYPE_CHECKING, Any, Protocol
 
-from raes_contracts.account_credentials import (
-    account_placement_has_credential_bindings,
-    value_free_account_placement_payload,
-)
 from raes_contracts.artifact_requirements import ArtifactSatisfactionDisclosureModel
 from raes_contracts.contracts import RealizationEnvelopeIdentityModel
 from raes_contracts.contracts.time_model import TimeRuntimeStateModel
@@ -30,6 +26,11 @@ from raes_contracts.runtime_state import (
 )
 
 from .control_plane_store_observations import realization_observation_from_payload
+from .control_plane_store_payloads import (
+    _entry_payloads,
+    _realization_observations_payload,
+    _realization_provenance_payload,
+)
 
 if TYPE_CHECKING:
     from .control_plane_store_local import LocalControlPlaneStore
@@ -177,18 +178,7 @@ def _snapshot_payload(snapshot: RuntimeSnapshot) -> dict[str, Any]:
     require_participant_autonomous_runtime_snapshot(snapshot)
     payload = {
         "schema_version": RuntimeSnapshotEnvelope().schema_version,
-        "entries": {
-            address: {
-                "address": entry.address,
-                "domain": entry.domain.value,
-                "resource_type": entry.resource_type,
-                "payload": dict(entry.payload),
-                "ordering_dependencies": list(entry.ordering_dependencies),
-                "refresh_dependencies": list(entry.refresh_dependencies),
-                "status": entry.status,
-            }
-            for address, entry in snapshot.entries.items()
-        },
+        "entries": _entry_payloads(snapshot),
         "orchestration_results": dict(snapshot.orchestration_results),
         "orchestration_history": {address: list(events) for address, events in snapshot.orchestration_history.items()},
         "evaluation_results": dict(snapshot.evaluation_results),
@@ -229,67 +219,13 @@ def _snapshot_payload(snapshot: RuntimeSnapshot) -> dict[str, Any]:
         "time_model_state": (
             snapshot.time_model_state.model_dump(mode="json") if snapshot.time_model_state is not None else None
         ),
-        "realization_provenance": [
-            {
-                "address": entry.address,
-                "field_path": entry.field_path,
-                "domain": entry.domain,
-                "requirement_kind": entry.requirement_kind,
-                "explicitness": entry.explicitness.value,
-                "provenance": entry.provenance.value,
-                "governing_scope": entry.governing_scope,
-                "artifact_satisfaction": (
-                    entry.artifact_satisfaction.model_dump(mode="json")
-                    if entry.artifact_satisfaction is not None
-                    else None
-                ),
-            }
-            for entry in snapshot.realization_provenance
-        ],
-        "realization_observations": [
-            {
-                "address": entry.address,
-                "field_path": entry.field_path,
-                "domain": entry.domain,
-                "requirement_kind": entry.requirement_kind,
-                "verification_scope": entry.verification_scope.value,
-                "observation_strength": entry.observation_strength.value,
-                **(
-                    {
-                        "observed_value": entry.observed_value,
-                        "operating_system": (
-                            {
-                                "family": entry.operating_system.family,
-                                "distribution": entry.operating_system.distribution,
-                                "version": entry.operating_system.version,
-                            }
-                            if entry.operating_system is not None
-                            else None
-                        ),
-                        "operation_id": entry.operation_id,
-                        "envelope_digest": entry.envelope_digest,
-                        "configuration_digest": entry.configuration_digest,
-                        "observer_version": entry.observer_version,
-                        "sequence": entry.sequence,
-                        "binding_verified": entry.binding_verified,
-                    }
-                    if entry.requirement_kind in {"compute-substrate", "operating-system"}
-                    else {}
-                ),
-            }
-            for entry in snapshot.realization_observations
-        ],
+        "realization_provenance": _realization_provenance_payload(snapshot),
+        "realization_observations": _realization_observations_payload(snapshot),
         "realization_envelope": (
             snapshot.realization_envelope.model_dump(mode="json") if snapshot.realization_envelope is not None else None
         ),
         "metadata": dict(snapshot.metadata),
     }
-    for entry in payload["entries"].values():
-        if entry["resource_type"] != "account-placement":
-            continue
-        entry_payload = entry["payload"]
-        if account_placement_has_credential_bindings(entry_payload):
-            entry["payload"] = value_free_account_placement_payload(entry_payload)
     return payload
